refactor(webview): split hotspot modules
